Bugfix in ExtensionServer.
Now that we are passing a second, `meta` arg to request handlers, we cannot use `browser.runtime.sendMessage` as a handler directly;
we need a wrapper that can forward just the first arg, and ignore `meta`.

Accompanying this change, I have reviewed all existing usages of request handlers in this repo, as well as in the pfsc-ext and pfsc-ise repos,
to look for similar bugs. I found none.